server/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-mysql==4.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.5.0
python-decouple==3.6
pytz==2022.5
sqlparse==0.4.3
mysqlclient==2.1.1
gunicorn==20.1.0
whitenoise==6.2.0